test: Move Tests into v1 folder
