+++ kuksa-java-sdk/src/test/kotlin/org/eclipse/kuksa/connectivity/databroker/v1/extensions/DataBrokerTransporterExtensions.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 Contributors to the Eclipse Foundation
+ * Copyright (c) 2023 - 2025 Contributors to the Eclipse Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,7 +17,7 @@
  *
  */
 
-package org.eclipse.kuksa.extensions
+package org.eclipse.kuksa.connectivity.databroker.v1.extensions
 
 import io.kotest.assertions.fail
 import org.eclipse.kuksa.connectivity.databroker.v1.DataBrokerTransporter
