+++ kuksa-java-sdk/src/test/kotlin/org/eclipse/kuksa/connectivity/databroker/v1/provider/DataBrokerConnectorProvider.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 Contributors to the Eclipse Foundation
+ * Copyright (c) 2023 - 2025 Contributors to the Eclipse Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,9 +14,10 @@
  * limitations under the License.
  *
  * SPDX-License-Identifier: Apache-2.0
+ *
  */
 
-package org.eclipse.kuksa.connectivity.databroker
+package org.eclipse.kuksa.connectivity.databroker.v1.provider
 
 import io.grpc.ChannelCredentials
 import io.grpc.Grpc
@@ -24,10 +25,13 @@ import io.grpc.ManagedChannel
 import io.grpc.ManagedChannelBuilder
 import io.grpc.TlsChannelCredentials
 import org.eclipse.kuksa.connectivity.authentication.JsonWebToken
-import org.eclipse.kuksa.connectivity.authentication.JwtType
+import org.eclipse.kuksa.connectivity.databroker.DATABROKER_HOST
+import org.eclipse.kuksa.connectivity.databroker.DATABROKER_TIMEOUT_SECONDS
+import org.eclipse.kuksa.connectivity.databroker.DATABROKER_TIMEOUT_UNIT
 import org.eclipse.kuksa.connectivity.databroker.docker.DEFAULT_PORT_INSECURE
 import org.eclipse.kuksa.connectivity.databroker.docker.DEFAULT_PORT_SECURE
 import org.eclipse.kuksa.connectivity.databroker.v1.DataBrokerConnector
+import org.eclipse.kuksa.mocking.JwtType
 import org.eclipse.kuksa.model.TimeoutConfig
 import org.eclipse.kuksa.test.TestResourceFile
 import java.io.IOException
@@ -35,6 +39,7 @@ import java.io.InputStream
 
 class DataBrokerConnectorProvider {
     lateinit var managedChannel: ManagedChannel
+
     fun createInsecure(
         host: String = DATABROKER_HOST,
         port: Int = DEFAULT_PORT_INSECURE,
