+++ kuksa-java-sdk/src/test/kotlin/org/eclipse/kuksa/mocking/JwtType.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 - 2024 Contributors to the Eclipse Foundation
+ * Copyright (c) 2023 - 2025 Contributors to the Eclipse Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,7 +17,7 @@
  *
  */
 
-package org.eclipse.kuksa.connectivity.authentication
+package org.eclipse.kuksa.mocking
 
 import org.eclipse.kuksa.test.TestResourceFile
 import java.io.InputStream
